Use mdoc-net.exe, not mdoc.exe.
'mdoc export-html' requires XSLT 2.0 support (specifically, for result tree
fragments to be usable as node-sets), while .NET's XSLT support is limited to
XSLT 1.0.  The only way to get mdoc usable under .NET, then, is to bundle
Mono's System.Xml.dll with mdoc, and since it's a GAC'd copy that's not
possible.

The result is a new program, 'mdoc-net.exe', which is:

	mdoc.exe + Mono's System.Xml.dll + monodoc.dll

Yes, it's bigger this way, but it actually works under .NET.