Tuples needs to generate multiple files.
CodeDom has no facility to do conditional compilation (#if FOO), so
we need to generate multiple files and let the solution sort out
whether to include the Tuples type or not, and place the Tuples extension
methods that we always want into a separate file.

Add '-o' parsing support to FileGenerator, and have Tuples.exe generate
multiple files when -o is given (e.g. Tuples.cs and TuplesCoda.cs).