Provide a simpler way to create Either<...> instances.
Thanks to Dunb on ##csharp@irc.freenode.net for the idea.

We add implicit conversion operators to Either<...> so that we can implicitly
convert values into an Either<...> instance:

	Either<int, double> v = 42.0;

Note that this will NOT work if all type parameters are the same (e.g. with
Either<double, double>), but for the common case in which things are different
(by far the most common case), it should greatly simplify things.