Fix the CreateDependent_ChangeStepUpdatesProgress() Test.
The problem (aside from me not checking that it compiled and worked?!) was
how NotifyingProperty.CreateDependent() works: it invokes the PropertyChanged
handler from within a PropertyChanged notification.

As a result, since the NotifyingProperty.CreateDependent() PropertyChanged
event handler comes before the test's PropertyChanged event handler (and thus
is invoked first), PropertyChanged notifications are effectively in reverse
order -- even though CurrentStep is the property that was changed (and thus is
triggered first), Progress is the first property that the test sees has
changed.  This may be counter to expectations.