Don't use the Condition attribute in <Compile/> sources.
The problem is that MonoDevelop doesn't have full MSBuild support, so if/when
it sees a <Compile Include="..." Condition="..." /> element, it *skips* the
file outright.  Result: tons of compilation errors, as Tuple<...> aren't
defined anywhere (surprise!).

We'll need to revisit "easily supporting .NET 4.0" in the future, when
MonoDevelop has better MSBuild support.