Remove ReaderWriterLockSlim dependency from Synchronized* types.
ReaderWriterLockSlim has platform dependencies (e.g. it doesn't exist in
MonoTouch and Silverlight profiles), and thus code using it can't be in
Cadenza.dll (as per current guidelines).

To remove the ReaderWriterLockSlim dependency, we instead require that
SynchronizedCollection<T> take two Func<Action> parameters.

The semantics are that invoking the Func<Action> will acquire the lock and
return an Action which, when invoked, will release the lock previously
acquired.

This allows SynchronizedCollection<T> to be used with both normal monitors
(via the new ObjectCoda.CreateMonitorLockAccessor() helper, which is the
default implementation) or with ReaderWriterLockSlim (via the new
ReaderWriterLockSlimCoda.CreateReadLockAccessor() and
ReaderWriterLockSlimCoda.CreateWriteLockAccessor() extension methods).