Terminology change: Parse is for strings, Convert for type changes.
Change Cadenza.Maybe and Cadenza.Either so that there is only a
TryParse(string) method (parsing is for strings), and rename the
TryParse(TSource) method to TryConvert(TSource) (convert is for type changes).

This terminology change is for clarity, and consistency with the BCL.
(Consider: the implementation of Either.TryConvert() uses a TypeConverter and
the TypeConverter.ConvertFrom() and TypeConverter.ConvertTo() methods.
Pre-existing terminology is "Convert", not "Parse".)

Add an Either.TryConvert<R>(object) method (and consider what
"TryParse(object)" would mean, another reason for the terminology change).