Fix MathContract<T>.Log() expected exceptions.
When testing ExpressionMath<int>, converting -infinity to int results in an
OverflowException (which MathContract<T>.Log() checks for), but since the
OverflowException is generated within Either.TryConvert(), the exception is
wrapped within a NotSupportedException.

Thus, a NotSupportedException containing an OverflowException is a
successfully supported type, while a NotSupportedException that doesn't wrap
OverflowException is not.