Fix ICollectionContract<T> tests.
For "sanity", ICollectionContract<T>.RunAllTests should run
CopyTo_Exceptions() before CopyTo(), so that we can assert proper error
behavior before we assert proper regular behavior.

Fix the ICollectionContract<T>.CopyTo_Exceptions() ArgumentOutOfRangeException
check so that we only test one error condition:

	Assert.Throws<ArgumentOutOfRangeException>(() => c.CopyTo (new T [1], -1));

has two bugs: the array is two small (`c` contains 3 items, and the created
array of 1 element can't hold 3 elements) and the index is invalid.  The
intention is to check for the invalid index, not the too-small array (which is
checked later in the same method).