Fix the MathTests.Default_ThrowsIfCadenzaCoreCannotBeFound() test on .NET.
When you load an Assembly into an AppDomain via AppDomain.Load(byte[]), that
assembly isn't "findable" via AssemblyName.  Thus, when trying to do:

	AppDomain.DoCallBack(() => {var m = Math<string>.Default;});

It would fail as it was attempting to load the Cadenza-Tests.dll assembly into
the new AppDomain, which it couldn't do (as assembly probing is disabled for
that AppDomain).

The workaround is to use Reflection to get the Type for MathTests that we
injected into the AppDomain, and create a delegate over that type+method.