Turn static fields into instance fields.
It's possible that the static constructor could throw an exception during
construction if e.g. `T` doesn't support conversion from/to Int32.  This would
be bad, as it would result in a TypeLoadException, which would prevent others
from e.g. using ExpressionMath<T> as a base class for their own use.

Thus, turn these fields into instance fields.  If an exception is generated,
it'll be from the instance constructor and NOT a TypeLoadException.