Check for division by zero in MathContract<T>.
Lots of seemingly unrelated changes, which is probably showing how integrated
things are getting...

The question: what should Math<double>.Default.Divide(1.0, 0.0) do?

The answer: whatever (1.0/0.0) does: return double.PositiveInfinity.

This simple question and answer thus requires changing lots of things...

To wit:
	- Math<T>.IsNaN, Math<T>.IsInfinite(), and Math<T>.IsIEEE() should NOT
	  throw NotSupportedException; instead, they return `false` by default.
	  The reason why is so that ExpressionMath<T>.QuotientRemainder() can
	  rely on them to check the return value, so that the result of
	  ExpressionMath<T>.divide() will only be converted through `int` if
	  the value is !IsNaN() and !IsInfinite().
	- ExpressionMath<T> now overrides Math<T>.IsNaN().
	- MathContract<T> gains additional tests for division by 0 behavior.
	- Default Math<T> implementations don't check for division by zero;
	  the appropriate exception will be thrown at runtime.
	- DoubleMath and FloatMath check for IsNaN()/etc. as appropriate.

and seemingly unrelated, Either.TryConvert`2() now no longer indiscriminately
wraps the exception in a NotSupportedException.  Instead, it checks the
exception type to see if it provides a (string, Exception) constructor.  If it
does, it creates a new instance of the same Exception type, wrapping the
original exception.  Otherwise, NotSupportedException is used (as before).

This was done because Math<T>.FromIConvertible() uses Either.TryConvert`2(),
and Math<T>.FromIConvertible() is used from Math<T>.Log(), Math<T>.Sqrt(),
etc., and is thus invoked by Int32Math.Sqrt() (as it doesn't override that
method, and instead uses the base Math<T>.Sqrt() implementation).

The result is that, due to the new division by zero tests, we were now
attempting to convert double.NaN to int, which throws OverflowException, which
is wrapped by Either.TryConvert`2(), which becomes NotSupportedException,
which just seemed...ugly.  Also, duplicative, as *sometimes*
NotSupportedException would be because T was an unsigned type, and *sometimes*
because it didn't support conversion from NaN, so we could assert less.